+++ packages/ui/src/components/DataStory/Form/StringableInput.tsx
